# openapi.yaml
openapi: 3.0.3
info:
  title: WhatsApp Document Sender API
  version: "1.0.0"
  description: API REST para enviar mensajes y documentos por WhatsApp. (Versión pública sin autenticación)
servers:
  - url: https://document-whatsapp-sender-api.onrender.com/
    description: URL pública (reemplaza por tu URL de Render)
paths:
  /connect/{clientId}:
    post:
      summary: Conectar cliente y generar QR
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estado de conexión o QR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectResponse'
    get:
      summary: Consultar estado de conexión
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estado de la sesión
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectResponse'
  /qr/{clientId}:
    get:
      summary: Obtener imagen PNG del QR
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Imagen PNG del QR
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: No hay QR disponible
          content:
            text/html:
              schema:
                type: string
  /send/{clientId}:
    post:
      summary: Enviar documento o mensaje de texto
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendRequest'
      responses:
        '200':
          description: Mensaje enviado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Parámetros inválidos o cliente no conectado
          content:
            application/json:
              schema:
                type: object
                properties:
                  estado:
                    type: string
                  mensaje:
                    type: string
  /status/{clientId}:
    get:
      summary: Estado simple de la sesión
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
components:
  schemas:
    SendRequest:
      type: object
      properties:
        telefono:
          type: string
          description: Número destino (ej. 50671234567)
        url_documento:
          type: string
          format: uri
          description: URL pública del documento (opcional)
        caption:
          type: string
          description: Texto o leyenda (opcional)
      required:
        - telefono
    SendResponse:
      type: object
      properties:
        estado:
          type: string
          example: enviado
        mensaje:
          type: string
        id_mensaje:
          type: string
        destinatario:
          type: string
    ConnectResponse:
      type: object
      properties:
        status:
          type: string
          example: needs-scan
        qr:
          type: string
          description: Data URL del QR (base64) o null
        qr_url:
          type: string
          description: URL pública para ver el QR
        message:
          type: string
    StatusResponse:
      type: object
      properties:
        clientId:
          type: string
        status:
          type: string
        connected:
          type: boolean